rename v1health to v1_health
